Fixed issue where push promise can race finishing response
