+++ pom.xml
@@ -44,7 +44,7 @@
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     <checkstyle.config.location>src/conf/checkstyle/configuration.xml</checkstyle.config.location>
     <checkstyle.suppressions.location>src/conf/checkstyle/suppressions.xml</checkstyle.suppressions.location>
-    <jacoco.coverage.ratio>0.60</jacoco.coverage.ratio>
+    <jacoco.coverage.ratio>0.58</jacoco.coverage.ratio>
     <jacoco.missed.count>45</jacoco.missed.count>
 
     <junit.version>4.12</junit.version>
