+++ src/main/java/org/reaktivity/nukleus/http_cache/internal/proxy/request/InitialRequest.java
@@ -15,198 +15,55 @@
  */
 package org.reaktivity.nukleus.http_cache.internal.proxy.request;
 
-import static org.reaktivity.nukleus.http_cache.internal.stream.util.HttpHeaders.ETAG;
-import static org.reaktivity.nukleus.http_cache.internal.stream.util.HttpHeadersUtil.getHeaderOrDefault;
-
 import java.util.function.LongSupplier;
 
-import org.agrona.MutableDirectBuffer;
 import org.reaktivity.nukleus.buffer.BufferPool;
 import org.reaktivity.nukleus.function.MessageConsumer;
-import org.reaktivity.nukleus.http_cache.internal.proxy.cache.Cache;
-import org.reaktivity.nukleus.http_cache.internal.proxy.request.Request.Type;
-import org.reaktivity.nukleus.http_cache.internal.stream.util.Slab;
-import org.reaktivity.nukleus.http_cache.internal.types.HttpHeaderFW;
-import org.reaktivity.nukleus.http_cache.internal.types.ListFW;
-import org.reaktivity.nukleus.http_cache.internal.types.OctetsFW;
-import org.reaktivity.nukleus.http_cache.internal.types.stream.DataFW;
-import org.reaktivity.nukleus.http_cache.internal.types.stream.EndFW;
 import org.reaktivity.nukleus.route.RouteManager;
 
 public class InitialRequest extends CacheableRequest
 {
-    final BufferPool responseBufferPool;
-    int responseSlot = Slab.NO_SLOT;
-    int responseHeadersSize;
-    int responseSize;
-    boolean cachingResponse;    // TODO, consider using state management via method references
-    final MessageConsumer connect;
-    final long connectRef;
-    final LongSupplier supplyCorrelationId;
-    final LongSupplier supplyStreamId;
 
     public InitialRequest(
-        String acceptName,
-        MessageConsumer acceptReply,
-        long acceptReplyStreamId,
-        long acceptCorrelationId,
-        MessageConsumer connect,
-        long connectRef,
-        LongSupplier supplyCorrelationId,
-        LongSupplier supplyStreamId,
-        int requestURLHash,
-        BufferPool responseBufferPool,
-        BufferPool requestBufferPool,
-        int requestSlot,
-        int requestSize,
-        RouteManager router,
-        short authScope,
-        String etag)
+            String acceptName,
+            MessageConsumer acceptReply,
+            long acceptReplyStreamId,
+            long acceptCorrelationId,
+            MessageConsumer connect,
+            long connectRef,
+            LongSupplier supplyCorrelationId,
+            LongSupplier supplyStreamId,
+            int requestURLHash,
+            BufferPool responseBufferPool,
+            BufferPool requestBufferPool,
+            int requestSlot,
+            int requestSize,
+            RouteManager router,
+            short authScope,
+            String etag)
     {
         super(acceptName,
               acceptReply,
               acceptReplyStreamId,
               acceptCorrelationId,
-              router,
+              connect,
+              connectRef,
+              supplyCorrelationId,
+              supplyStreamId,
+              requestURLHash,
+              responseBufferPool,
               requestBufferPool,
               requestSlot,
               requestSize,
-              requestURLHash,
+              router,
               authScope,
               etag);
-        this.responseBufferPool = responseBufferPool;
-        this.cachingResponse = true;
-        this.supplyCorrelationId = supplyCorrelationId;
-        this.supplyStreamId = supplyStreamId;
-
-        this.connect = connect;
-        this.connectRef = connectRef;
     }
 
     @Override
     public Type getType()
     {
-        return Type.CACHEABLE;
+        return Type.INITIAL_REQUEST;
     }
 
-    // TODO remove need for duplication
-    public void copyRequestTo(MutableDirectBuffer buffer)
-    {
-        MutableDirectBuffer requestBuffer = requestBufferPool().buffer(requestSlot());
-        requestBuffer.getBytes(0, buffer, 0, requestSize());
-    }
-
-    public void cache(
-            ListFW<HttpHeaderFW> responseHeaders,
-            Cache cache)
-    {
-        etag(getHeaderOrDefault(responseHeaders, ETAG, etag()));
-
-        setupResponseBuffer();
-        MutableDirectBuffer buffer = responseBuffer();
-        final int headersSize = responseHeaders.sizeof();
-        buffer.putBytes(responseSize, responseHeaders.buffer(), responseHeaders.offset(), headersSize);
-        responseSize += headersSize;
-        this.responseHeadersSize = headersSize;
-    }
-
-    private void setupResponseBuffer()
-    {
-        this.responseSlot = responseBufferPool.acquire(acceptReplyStreamId());
-        this.responseHeadersSize = 0;
-        this.responseSize = 0;
-    }
-
-    public void cache(DataFW data)
-    {
-        if (cachingResponse)
-        {
-            OctetsFW payload = data.payload();
-            int sizeof = payload.sizeof();
-            if (responseSize + sizeof > responseBufferPool.slotCapacity())
-            {
-                this.purge();
-            }
-            else
-            {
-                MutableDirectBuffer buffer = responseBuffer();
-                buffer.putBytes(responseSize, payload.buffer(), payload.offset(), sizeof);
-                responseSize += sizeof;
-            }
-        }
-    }
-
-    public void cache(EndFW end, Cache cache)
-    {
-        if (cachingResponse)
-        {
-            cache.put(requestURLHash(), this);
-        }
-    }
-
-    public void purge()
-    {
-        super.purge();
-        if (responseSlot != Slab.NO_SLOT)
-        {
-            responseBufferPool.release(responseSlot);
-        }
-        this.responseSlot = Slab.NO_SLOT;
-        this.cachingResponse = false;
-    }
-
-    public long connectRef()
-    {
-        return connectRef;
-    }
-
-    public LongSupplier supplyCorrelationId()
-    {
-        return supplyCorrelationId;
-    }
-
-    public LongSupplier supplyStreamId()
-    {
-        return supplyStreamId;
-    }
-
-    // TODO hide abstraction
-    public int responseSlot()
-    {
-        return responseSlot;
-    }
-
-    // TODO hide abstraction
-    public int responseHeadersSize()
-    {
-        return responseHeadersSize;
-    }
-
-    // TODO hide abstraction
-    public int responseSize()
-    {
-        return responseSize;
-    }
-
-    public ListFW<HttpHeaderFW> getResponseHeaders(
-        ListFW<HttpHeaderFW> responseHeadersRO)
-    {
-        MutableDirectBuffer responseBuffer = responseBuffer();
-        return responseHeadersRO.wrap(responseBuffer, 0, responseHeadersSize);
-    }
-
-    public MessageConsumer connect()
-    {
-        return connect;
-    }
-
-    private MutableDirectBuffer responseBuffer()
-    {
-        return responseBufferPool.buffer(responseSlot);
-    }
-
-    public MutableDirectBuffer getData(BufferPool bp)
-    {
-        return bp.buffer(responseSlot);
-    }
 }
